Move difficulty attributes to API
