+++ app/rulesets/osu/performance/pp241007/pp.go
@@ -2,6 +2,7 @@ package pp241007
 
 import (
 	"github.com/wieku/danser-go/app/beatmap/difficulty"
+	"github.com/wieku/danser-go/app/rulesets/osu/performance/api"
 	"github.com/wieku/danser-go/app/rulesets/osu/performance/pp241007/skills"
 	"github.com/wieku/danser-go/framework/math/mutils"
 	"math"
@@ -22,7 +23,7 @@ type PPv2Results struct {
 type PPv2 struct {
 	Results PPv2Results
 
-	attribs Attributes
+	attribs api.Attributes
 
 	experimental bool
 
@@ -40,7 +41,7 @@ type PPv2 struct {
 	amountHitObjectsWithAccuracy int
 }
 
-func (pp *PPv2) PPv2x(attribs Attributes, combo, n300, n100, n50, nmiss int, diff *difficulty.Difficulty) PPv2 {
+func (pp *PPv2) PPv2x(attribs api.Attributes, combo, n300, n100, n50, nmiss int, diff *difficulty.Difficulty) PPv2 {
 	attribs.MaxCombo = max(1, attribs.MaxCombo)
 
 	if combo < 0 {
